added tests, moved Type enum in room out to a separate class
