+++ questmaster/src/main/java/pdc/Type.java
@@ -1,0 +1,3 @@
+package pdc;
+
+public enum Type {OPAQUE,TRANSPARENT,ARCHWAY,DOOR}
